PoS: move storage lazy collection handles into sub-mod
