+++ apps/src/lib/node/ledger/shell/prepare_proposal.rs
@@ -514,9 +514,9 @@ mod test_prepare_proposal {
     use namada::ledger::gas::Gas;
     use namada::ledger::pos::PosQueries;
     use namada::ledger::replay_protection;
+    use namada::proof_of_stake::storage::consensus_validator_set_handle;
     use namada::proof_of_stake::types::WeightedValidator;
     use namada::proof_of_stake::{
-        consensus_validator_set_handle,
         read_consensus_validator_set_addresses_with_stake, Epoch,
     };
     #[cfg(feature = "abcipp")]
