+++ apps/src/lib/node/ledger/shell/vote_extensions/bridge_pool_vext.rs
@@ -277,13 +277,13 @@ mod test_bp_vote_extensions {
     use namada::ledger::eth_bridge::EthBridgeQueries;
     use namada::ledger::pos::PosQueries;
     use namada::ledger::storage_api::StorageWrite;
+    use namada::proof_of_stake::storage::consensus_validator_set_handle;
     use namada::proof_of_stake::types::{
         Position as ValidatorPosition, WeightedValidator,
     };
     use namada::proof_of_stake::{
-        become_validator, consensus_validator_set_handle,
-        read_consensus_validator_set_addresses_with_stake, BecomeValidator,
-        Epoch,
+        become_validator, read_consensus_validator_set_addresses_with_stake,
+        BecomeValidator, Epoch,
     };
     use namada::proto::{SignableEthMessage, Signed};
     use namada::tendermint_proto::abci::VoteInfo;
