+++ apps/src/lib/node/ledger/shell/vote_extensions/eth_events.rs
@@ -454,11 +454,9 @@ mod test_vote_extensions {
     };
     use namada::eth_bridge::storage::bridge_pool;
     use namada::ledger::pos::PosQueries;
+    use namada::proof_of_stake::read_consensus_validator_set_addresses_with_stake;
+    use namada::proof_of_stake::storage::consensus_validator_set_handle;
     use namada::proof_of_stake::types::WeightedValidator;
-    use namada::proof_of_stake::{
-        consensus_validator_set_handle,
-        read_consensus_validator_set_addresses_with_stake,
-    };
     #[cfg(feature = "abcipp")]
     use namada::proto::{SignableEthMessage, Signed};
     use namada::tendermint_proto::abci::VoteInfo;
