+++ wasm/wasm_source/src/tx_withdraw.rs
@@ -21,8 +21,8 @@ fn apply_tx(ctx: &mut Ctx, tx_data: Tx) -> TxResult {
 #[cfg(test)]
 mod tests {
     use namada::ledger::pos::{PosParams, PosVP};
+    use namada::proof_of_stake::storage::unbond_handle;
     use namada::proof_of_stake::types::GenesisValidator;
-    use namada::proof_of_stake::unbond_handle;
     use namada::types::dec::Dec;
     use namada::types::storage::Epoch;
     use namada_tests::log::test;
